Change test data for granule collector test to not require axis order system property
